+++ ready/AzNamingTool/Helpers/Config.cs
@@ -6,5 +6,6 @@
         public string? AdminPassword { get; set; }
         public string? APIKey { get; set; }
         public string? AppTheme { get; set; }
+        public bool? DevMode{ get; set; } = false;
     }
 }