+++ ready/AzNamingTool/VERSIONHISTORY.md
@@ -4,10 +4,25 @@
 
 <img src="./wwwroot/images/AzureNamingToolLogo.png?raw=true" alt="Azure Naming Tool" title="Azure Naming Tool" height="150"/>
 
-## Version 2.3.1 (current)
+## Version 2.3.2 (current)
 
 ### Bug fixes
-- Updated Configuration page to export Admin configuration, if selecteddocumentation
+- Updated Reference/Generate pages to reload configuration data, if empty
+- Updated site to exclude Custom Component if no options are entered
+- Updated resourcetypes.json regex patterns
+- Updated cache reload process to ensure all data is loaded
+- Updated Admin Page styling
+- Updated comments in API
+- Updated Reference page to include delimiter for samples.
+- Added warning to Generate page for resource types with a lower-level scope.
+- Updated Nuget packages
+
+***
+
+## Version 2.3.1
+
+### Bug fixes
+- Updated Configuration page to export Admin configuration, if selected
 - Updated ResourceTypes.cs model to not require Short Name.
 
 ***